change the time type and add some data

+++ backend/src/main/java/cookme/recipesmodel/Recipes.java
@@ -7,7 +7,7 @@ import java.util.List;
 public record Recipes(@With String id,
                       String name,
                       String description,
-                      Integer time,
+                      double time,
                       String imageUrl,
                       String preparation,
                       Favorite status,
